Re-engineering: We divide the checker in two stages. When checking a program, it is not necessary that it results in an executable as in the case of libraries. The programer might just want to check for correctness.
But if we want an executable, we require that there be a main function and that this function does not depend
on type instances that in turn depend on abstract type instances.
So in the first stage, we make sure that each type and associated constructors are valid, each constructor invocation resolves to a valid constructor, each function call has at least one function declaration that may be called, statements are correct
and each operator has the associated function present.

In stage 2, we require that there be a main function, and that each constructor invocation resolve to a type instance without
dependence to abstract type instances  and function call resolve to a unique function and do not depend on abstract type instances as well.